fix(security+validation): callbackUrl XSS/open-redirect, deactivated redirect loop, date/id schemas
SECURITY:
- login-form: callbackUrl was pushed to router.push unsanitized -> javascript:
  URL = XSS, https://evil = open redirect. Now routed through safeInternalPath().
- proxy.ts: deactivated user (token.isActive=false) caused infinite redirect loop
  (layout sent to /login, proxy bounced back to /). isLoggedIn now treats
  isActive===false as not-logged-in for page routing.

VALIDATION (Invalid input -> Prisma 500):
- finance.schemas requiredDate/optionalDate: z.string().transform(new Date) accepted
  unparseable strings -> Invalid Date persisted. Now z.coerce.date().
- sales/crm/manufacturing id schemas: added .int() (fractional id reached Prisma Int).

+++ src/lib/validations/manufacturing.schemas.ts
@@ -30,7 +30,7 @@ export type UpdateProductInput = z.infer<typeof updateProductSchema>
 // ==================== PRODUCTION ORDER ====================
 
 export const createProductionOrderSchema = z.object({
-  productId: z.coerce.number().min(1, "Produk wajib dipilih"),
+  productId: z.coerce.number().int().min(1, "Produk wajib dipilih"),
   qty: z.coerce.number().min(1, "Qty minimal 1"),
   startDate: optionalString(30),
   endDate: optionalString(30),
@@ -40,7 +40,7 @@ export const createProductionOrderSchema = z.object({
 export type CreateProductionOrderInput = z.infer<typeof createProductionOrderSchema>
 
 export const updateProductionOrderSchema = z.object({
-  productId: z.coerce.number().min(1, "Produk wajib dipilih"),
+  productId: z.coerce.number().int().min(1, "Produk wajib dipilih"),
   qty: z.coerce.number().min(1, "Qty minimal 1"),
   startDate: optionalString(30),
   endDate: optionalString(30),
